Defined class structure, wrote TODO's, implemented decoding framework
